+++ src/main/java/pl/otros/intellij/JumpToCode/logic/SourceFile.java
@@ -3,11 +3,11 @@ package pl.otros.intellij.JumpToCode.logic;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 
-class SourceFile {
+public class SourceFile {
   VirtualFile virtualFile;
   private Project project;
 
-  SourceFile(Project project, VirtualFile virtualFile) {
+  public SourceFile(Project project, VirtualFile virtualFile) {
     this.project = project;
     this.virtualFile = virtualFile;
   }
